add zeknox build tag
