--- backend/groth16/bn254/zeknox/provingkey_nozeknox.go
@@ -1,41 +1,0 @@
-//go:build !zeknox
-
-package zeknox_bn254
-
-import (
-	"unsafe"
-
-	groth16_bn254 "github.com/consensys/gnark/backend/groth16/bn254"
-	cs "github.com/consensys/gnark/constraint/bn254"
-)
-
-type deviceInfo struct {
-	G1Device struct {
-		A, B, K, Z unsafe.Pointer
-	}
-	DomainDevice struct {
-		Twiddles, TwiddlesInv     unsafe.Pointer
-		CosetTable, CosetTableInv unsafe.Pointer
-	}
-	G2Device struct {
-		B unsafe.Pointer
-	}
-	DenDevice             unsafe.Pointer
-	InfinityPointIndicesK []int
-}
-
-type ProvingKey struct {
-	groth16_bn254.ProvingKey
-	*deviceInfo
-}
-
-func Setup(r1cs *cs.R1CS, pk *ProvingKey, vk *groth16_bn254.VerifyingKey) error {
-	return groth16_bn254.Setup(r1cs, &pk.ProvingKey, vk)
-}
-
-func DummySetup(r1cs *cs.R1CS, pk *ProvingKey) error {
-	return groth16_bn254.DummySetup(r1cs, &pk.ProvingKey)
-}
-
-func (pk *ProvingKey) Free() {
-}
